Alteração de estrutura de types.
