+++ models/ResponseBodySpotify.go
@@ -1,0 +1,9 @@
+package models
+
+type ResponseBody struct {
+	AccessToken  string `json:"access_token"`
+	TokenType    string `json:"token_type"`
+	Scope        string `json:"scope"`
+	ExpiresIn    int    `json:"expires_in"`
+	RefreshToken string `json:"refresh_token"`
+}
