--- models/Spotify/Track.go
@@ -1,7 +1,0 @@
-package models
-
-type Track struct {
-	ExternalURLs struct {
-		Spotify string `json:"spotify"`
-	} `json:"external_urls"`
-}
