+++ models/Track.go
@@ -1,0 +1,14 @@
+package models
+
+type TrackItem struct {
+	Track struct {
+		Name         string `json:"name"`
+		ExternalURLs struct {
+			Spotify string `json:"spotify"`
+		} `json:"external_urls"`
+	} `json:"track"`
+}
+
+type TrackResponse struct {
+	Items []TrackItem `json:"items"`
+}
